# backend/internal/domain/user.go
package domain

import "time"

// User is both the API user representation and the GORM model for users.
// PasswordHash is deliberately never serialized into an API response.
type User struct {
	ID           string    `gorm:"primaryKey;type:uuid" json:"id"`
	Email        string    `gorm:"uniqueIndex;not null" json:"email"`
	Name         string    `gorm:"not null" json:"name"`
	PasswordHash string    `gorm:"not null" json:"-"`
	CreatedAt    time.Time `json:"createdAt"`
	UpdatedAt    time.Time `json:"updatedAt"`
}

type Registration struct {
	Email    string `json:"email"`
	Password string `json:"password"`
	Name     string `json:"name"`
}

type Login struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}