# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Composable SaaS API
  version: 1.0.0
  description: Provider-neutral API contract for generated SaaS applications.
servers:
  - url: /api/v1
paths:
...
  /users/me:
    get:
      operationId: currentUser
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: Current user, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
    delete:
      operationId: deleteAccount
      security: [{ sessionCookie: [] }]
      responses: { '204': { description: Account deleted }, '401': { $ref: '#/components/responses/Unauthorized' } }
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: session }
  responses:
...
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
...
    Error: { type: object, required: [error, code], properties: { error: { type: string }, code: { type: string } } }
...
    User: { type: object, required: [id, email, name], properties: { id: { type: string }, email: { type: string, format: email }, name: { type: string } } }
...